refactor: 카드 수정 모달 리팩토링
