+++ src/components/organism/CardEditModal/CardEditModal.stories.jsx
@@ -1,17 +1,20 @@
 import {
-  Button,
+  // Button,
   CardEditModal,
-  HashTagInput,
-  EmojiPickerActiveButton,
+  // HashTagInput,
+  // EmojiPickerActiveButton,
 } from '@components';
-import { useState } from 'react';
+// import { useState } from 'react';
 
 export default {
   title: 'Component/Domain/CardEditModal',
 };
 
 export const Default = args => {
-  return <CardEditModal {...args} />;
+  const handleSubmit = waffleCard => {
+    console.log('제출!', waffleCard);
+  };
+  return <CardEditModal {...args} onSubmit={handleSubmit} />;
 };
 Default.argTypes = {
   visible: {
@@ -26,46 +29,46 @@ Default.argTypes = {
   onSubmit: { action: 'onSubmit' },
 };
 
-export const Usage = () => {
-  const [visible, setVisible] = useState(false);
+// export const Usage = () => {
+//   const [visible, setVisible] = useState(false);
 
-  const handleClose = () => {
-    setVisible(false);
-  };
+//   const handleClose = () => {
+//     setVisible(false);
+//   };
 
-  return (
-    <>
-      <button type="button" onClick={() => setVisible(true)}>
-        Show CardEditModal
-      </button>
-      <CardEditModal visible={visible} onClose={handleClose} />
-    </>
-  );
-};
+//   return (
+//     <>
+//       <button type="button" onClick={() => setVisible(true)}>
+//         Show CardEditModal
+//       </button>
+//       <CardEditModal visible={visible} onClose={handleClose} />
+//     </>
+//   );
+// };
 
-export const Inputs = args => {
-  return <HashTagInput {...args} />;
-};
-Inputs.argTypes = {
-  onChange: { action: 'onChange' },
-};
+// export const Inputs = args => {
+//   return <HashTagInput {...args} />;
+// };
+// Inputs.argTypes = {
+//   onChange: { action: 'onChange' },
+// };
 
-export const PickerButton = args => {
-  return (
-    <>
-      <h4>이모지</h4>
-      <EmojiPickerActiveButton {...args} />
-      <h4>배경색</h4>
-      <Button backgroundColor="royalblue" />
-      <h4>해시태그</h4>
-      <Button backgroundColor="royalblue" />
-    </>
-  );
-};
-PickerButton.argTypes = {
-  disabled: {
-    defaultValue: false,
-    control: { type: 'boolean' },
-  },
-  onChange: { action: 'onChange' },
-};
+// export const PickerButton = args => {
+//   return (
+//     <>
+//       <h4>이모지</h4>
+//       <EmojiPickerActiveButton {...args} />
+//       <h4>배경색</h4>
+//       <Button backgroundColor="royalblue" />
+//       <h4>해시태그</h4>
+//       <Button backgroundColor="royalblue" />
+//     </>
+//   );
+// };
+// PickerButton.argTypes = {
+//   disabled: {
+//     defaultValue: false,
+//     control: { type: 'boolean' },
+//   },
+//   onChange: { action: 'onChange' },
+// };
